refactor: organize them into movie package
